Hide epilogue away in users and show how to use it alongside custom routes.
